backend/requirements.txt
fastapi==0.128.8
sqlalchemy==2.0.46
alembic==1.16.5
pydantic==2.12.5
pydantic-settings==2.2.1
pydantic[email]
psycopg[binary]==3.3.2
uvicorn==0.39.0
passlib==1.7.4
PyJWT==2.8.0
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
bcrypt==4.0.1
passlib[bcrypt]
